refactored tests
more unit tests. Some are not working. See TestFeatureConfig
