+++ src/Standard.Data.Json.Tests/Source/ClassTests.cs
@@ -11,6 +11,8 @@ using Xunit;
 
 namespace Standard.Data.Json.Tests
 {
+#if FEAT_SERIALIZE_INTERFACE
+
 	public class InterfaceTests
 	{
 		[Fact]
@@ -28,6 +30,8 @@ namespace Standard.Data.Json.Tests
 		}
 	}
 
+#endif
+
 	public class AbstractClassTests
 	{
 		[Fact]
@@ -37,6 +41,50 @@ namespace Standard.Data.Json.Tests
 			var json = JsonConvert.Serialize(p1);
 
 			Assert.True(!string.IsNullOrEmpty(json));
+		}
+	}
+
+	public class AccessTests
+	{
+		private class MyPrivateClass
+		{
+			public int ID { get; set; }
+			public string Name { get; set; }
+			public MyPrivateClass Inner { get; set; }
+		}
+
+		[Fact]
+		public void TestSerializeDeserializeNonPublicType()
+		{
+			string s;
+			var e = new List<E> { new E { V = 1 }, new E { V = 2 } };
+			s = JsonConvert.Serialize(e);
+			JsonConvert.Serialize(JsonConvert.Deserialize<List<E>>(s = JsonConvert.Serialize(e)));
+			Assert.True(!string.IsNullOrWhiteSpace(s));
+		}
+
+		[Fact]
+		public void SerializeNonPublicType()
+		{
+			var test = new MyPrivateClass { ID = 100, Name = "Test", Inner = new MyPrivateClass { ID = 200, Name = "Inner" } };
+			var json = JsonConvert.Serialize(test);
+			var data = JsonConvert.Deserialize<MyPrivateClass>(json);
+			Assert.True(json != null);
+		}
+
+		[Fact]
+		public void TestSerializeDeserializeNonPublicSetter()
+		{
+			var model = new Person("John", 12);
+
+			var json = JsonConvert.Serialize(model);
+
+			//var settings = new JsonSerializerSettings { IncludeTypeInformation = true };
+			JsonConvert.IncludeTypeInfo = true;
+			var deserializedModel = JsonConvert.Deserialize<Person>(json);
+			Assert.Equal("John", deserializedModel.Name);
+			Assert.Equal(12, deserializedModel.Age);
+			JsonConvert.IncludeTypeInfo = false;
 		}
 	}
