+++ src/Standard.Data.Json.Tests/Source/TestFeatureConfig.cs
@@ -1,10 +1,13 @@
 // this currently doesn't work in release mode
 // Test lives in SerializePolyTests.cs
-//
 //#define FEAT_POLY_SERIALIZATION
 
 // serialization of structs is untested
 // Test lives in SerializeStructTests.cs
-//
 //#define FEAT_STRUCT_SUPPORT
 
+// Test lives in ClassTests.cs
+//#define FEAT_SERIALIZE_INTERFACE
+
+// Test lives in XmlTests.cs
+//#define FEAT_SUPPORT_XML_ATTRIBUTE
